v0.01.22+++
單據同步:添加 命令類單據_單據結單狀態
單據同步:添加 命令類單據_固定180日移除資料
單據同步:添加 領料類單據_固定180日移除資料
單據同步:添加 入料類單據_固定180日移除資料
單據同步:修正 如果移除項目30日有效
System Cloud:修正 例外判斷
System Cloud:LOG 配置200G
System Cloud:BUG 權限S5/S6失效問題
BIOS:添加 BIOS版本管理50%
配料清單:修正 顯示排序內容
單據分配:修正 打印結構排序
單據分配:添加 產線配料清點人
單據分配:添加 產線PASS機制

+++ dtriCloud/service-client-12001/src/main/java/dtri/com/tw/login/CutomerUserDetailsService.java
@@ -60,7 +60,7 @@ public class CutomerUserDetailsService implements UserDetailsService {
 				if (ch[i] == '1') {
 					int move_p = 11 - i;
 					double now_p = Math.pow(10, move_p);
-					String now_s = String.format("%012d", (int) now_p);
+					String now_s = String.format("%012d", (long) now_p);
 					String role = systemGroup.getSystemPermission().getSpcontrol().replaceAll("\\.", "_") //
 							+ "_"//
 							+ now_s;
